Some more tidy up.
